Method to Modify Subsets of an Object By Reference

This is an S3 Method to replace or transform a subset of an object BY REFERENCE.

Usage

sb_set(x, ...)

# S3 method for default
sb_set(x, i, ..., rp, tf)

# S3 method for matrix
sb_set(x, row = NULL, col = NULL, i = NULL, ..., rp, tf)

# S3 method for array
sb_set(x, idx = NULL, dims = NULL, rcl = NULL, i = NULL, ..., rp, tf)

# S3 method for list
sb_set(x, i, ..., rp, tf)

# S3 method for data.frame
sb_set(x, row = NULL, col = NULL, filter = NULL, vars = NULL, ..., rp, tf)

Arguments

x

see subsets_classes.

...

further arguments passed to or from other methods.

i, row, col, idx, dims, rcl, filter, vars

See subsets_indx_args.
An empty index selection returns the original object unchanged.

rp

an object of somewhat the same type as the selected subset of x, and the same same length as the selected subset of x or a length of 1.

tf

the transformation function.

Value

Returns: VOID. This method modifies the object by REFERENCE.

Do NOT use assignment like x <- sb_set(x, ...).

Since this function returns void, you'll just get NULL.

Details

Transform or Replace
Specifying argument tf will transform the subset. Specifying rp will replace the subset. One cannot specify both tf and rp. It's either one set or the other.
Note that there is not sb_set() method for factors: this is intentional.

Warning

Due to the way replacement or transformation by reference works, types (see typeof) CANNOT be coerced to another type. Thus, for example, the following code:

x <- 1:16
sb_set(x, i = 1:8, rp = 8.5)
x

gives c(rep(8, 8) 9:16) instead of c(rep(8.5, 8), 9:16), because x is of type integer, so rp is interpreted as type integer also.

Examples

gen_mat <- function() {
  obj <- matrix(1:16, ncol = 4)
  colnames(obj) <- c("a", "b", "c", "a")
  return(obj)
}

# atomic objects ====

obj <- obj2 <- gen_mat()
obj
#>      a b  c  a
#> [1,] 1 5  9 13
#> [2,] 2 6 10 14
#> [3,] 3 7 11 15
#> [4,] 4 8 12 16
sb_set(obj, 1:3, 1:3, rp = -1:-9)
obj2
#>       a  b  c  a
#> [1,] -1 -4 -7 13
#> [2,] -2 -5 -8 14
#> [3,] -3 -6 -9 15
#> [4,]  4  8 12 16
obj <- obj2 <- gen_mat()
obj
#>      a b  c  a
#> [1,] 1 5  9 13
#> [2,] 2 6 10 14
#> [3,] 3 7 11 15
#> [4,] 4 8 12 16
sb_set(obj, i = \(x)x<=5, rp = -1:-5)
obj2
#>       a  b  c  a
#> [1,] -1 -5  9 13
#> [2,] -2  6 10 14
#> [3,] -3  7 11 15
#> [4,] -4  8 12 16
obj <- obj2 <- gen_mat()
obj
#>      a b  c  a
#> [1,] 1 5  9 13
#> [2,] 2 6 10 14
#> [3,] 3 7 11 15
#> [4,] 4 8 12 16
sb_set(obj, col = "a", rp = cbind(-1:-4, -5:-8))
obj2
#>       a b  c  a
#> [1,] -1 5  9 -5
#> [2,] -2 6 10 -6
#> [3,] -3 7 11 -7
#> [4,] -4 8 12 -8

obj <- obj2 <- gen_mat()
obj
#>      a b  c  a
#> [1,] 1 5  9 13
#> [2,] 2 6 10 14
#> [3,] 3 7 11 15
#> [4,] 4 8 12 16
sb_set(obj, 1:3, 1:3, tf = \(x) -x)
obj2
#>       a  b   c  a
#> [1,] -1 -5  -9 13
#> [2,] -2 -6 -10 14
#> [3,] -3 -7 -11 15
#> [4,]  4  8  12 16
obj <- obj2 <- gen_mat()
obj
#>      a b  c  a
#> [1,] 1 5  9 13
#> [2,] 2 6 10 14
#> [3,] 3 7 11 15
#> [4,] 4 8 12 16
sb_set(obj, i = \(x)x<=5, tf = \(x) -x)
obj2
#>       a  b  c  a
#> [1,] -1 -5  9 13
#> [2,] -2  6 10 14
#> [3,] -3  7 11 15
#> [4,] -4  8 12 16
obj <- obj2 <- gen_mat()
obj
#>      a b  c  a
#> [1,] 1 5  9 13
#> [2,] 2 6 10 14
#> [3,] 3 7 11 15
#> [4,] 4 8 12 16
sb_set(obj, col = "a", tf = \(x) -x)
obj2
#>       a b  c   a
#> [1,] -1 5  9 -13
#> [2,] -2 6 10 -14
#> [3,] -3 7 11 -15
#> [4,] -4 8 12 -16

gen_array <- function() {
  array(1:64, c(4,4,3))
}
obj <- gen_array()
obj
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    5    9   13
#> [2,]    2    6   10   14
#> [3,]    3    7   11   15
#> [4,]    4    8   12   16
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   17   21   25   29
#> [2,]   18   22   26   30
#> [3,]   19   23   27   31
#> [4,]   20   24   28   32
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   33   37   41   45
#> [2,]   34   38   42   46
#> [3,]   35   39   43   47
#> [4,]   36   40   44   48
#> 
sb_set(obj, list(1:3, 1:2, c(1, 3)), 1:3, rp = -1:-12)
obj
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   -1   -4    9   13
#> [2,]   -2   -5   10   14
#> [3,]   -3   -6   11   15
#> [4,]    4    8   12   16
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   17   21   25   29
#> [2,]   18   22   26   30
#> [3,]   19   23   27   31
#> [4,]   20   24   28   32
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   -7  -10   41   45
#> [2,]   -8  -11   42   46
#> [3,]   -9  -12   43   47
#> [4,]   36   40   44   48
#> 
obj <- gen_array()
obj
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    5    9   13
#> [2,]    2    6   10   14
#> [3,]    3    7   11   15
#> [4,]    4    8   12   16
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   17   21   25   29
#> [2,]   18   22   26   30
#> [3,]   19   23   27   31
#> [4,]   20   24   28   32
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   33   37   41   45
#> [2,]   34   38   42   46
#> [3,]   35   39   43   47
#> [4,]   36   40   44   48
#> 
sb_set(obj, i = \(x)x<=5, rp = -1:-5)
obj
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   -1   -5    9   13
#> [2,]   -2    6   10   14
#> [3,]   -3    7   11   15
#> [4,]   -4    8   12   16
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   17   21   25   29
#> [2,]   18   22   26   30
#> [3,]   19   23   27   31
#> [4,]   20   24   28   32
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   33   37   41   45
#> [2,]   34   38   42   46
#> [3,]   35   39   43   47
#> [4,]   36   40   44   48
#>